cv-analyzer-ai/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.115.2
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database and ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Document processing
pypdf2==3.0.1
pdfplumber==0.11.7
python-docx==1.1.2

# Natural Language Processing and ML
spacy==3.8.2
scikit-learn==1.6.1
pandas==2.3.2
numpy==2.3.2
nltk==3.9.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1

# HTTP and async support
httpx==0.28.1
aiofiles==24.1.0

# Validation and serialization
pydantic==2.10.3
pydantic-settings==2.6.1

# Logging and monitoring
structlog==24.4.0

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Caching
redis==5.2.1

# Date and time utilities
python-dateutil==2.9.0.post0